# shared/api_spec.yaml
openapi: 3.0.3
info:
  title: DPS Dashboard API
  description: >
    Deshapriya Park Durga Puja Club Management Dashboard API.
    All endpoints require authentication unless marked as public.
    Role enforcement is server-side on every route.
  version: 1.0.0
  contact:
    name: DPS Dashboard Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: http://{vps_ip}/api
    description: LunaNode VPS
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: HTTP-only JWT session cookie set by NextAuth
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
...
security:
  - cookieAuth: []
paths:
...
  /webhooks/razorpay:
    post:
      tags: [Webhooks]
      summary: Razorpay webhook handler
      description: >
        Receives payment.captured events from Razorpay.
        Verifies HMAC signature, creates full Transaction record,
        auto-approves, updates membership status, generates receipt,
        writes audit + activity logs, sends WhatsApp notifications.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '400':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== TRANSACTIONS ====================
...